Added comment lines to describe objective of Google signup implementation files and stored them in new folder.
