--- src/components/Authentication/Login.tsx
@@ -1,10 +1,0 @@
-import React from 'react'
-
-const Login = () => {
-  return (
-    <div>Login</div>
-
-  )
-}
-
-export default Login
