--- src/components/Authentication/Signup.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const Signup = () => { 
-    return (
-    <div>Signup</div>
-  )
-}
-
-export default Signup
