+++ src/components/Authentication/google_signup_unfinished/Login.tsx
@@ -1,0 +1,12 @@
+// Script for creating a button for logging into Firebase using Google.
+
+import React from 'react'
+
+const Login = () => {
+  return (
+    <div>Login</div>
+
+  )
+}
+
+export default Login
