+++ src/components/Authentication/google_signup_unfinished/Signup.tsx
@@ -1,0 +1,11 @@
+// Script for creating a button for signing into Firebase using Google.
+
+import React from 'react'
+
+const Signup = () => { 
+    return (
+    <div>Signup</div>
+  )
+}
+
+export default Signup
